moved styles folder outside of components folder
